+++ packages/next-shell/src/index.ts
@@ -1,3 +1,5 @@
 // Use this file to export React client components (e.g. those with 'use client' directive) or other non-server utilities
 import '@/tailwind/tailwind.css';
+export * from './lib/client-components';
 export * from './lib/providers';
+