+++ storybook/react/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/docs-templates';
+export * from './lib/decorators';