+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+## 0.28.6 (2024-04-07)
+
+
+### 🚀 Features
+
+- **next-storybook:** add aboilerplate nextjs project to house our storybook ([87ed7ea](https://github.com/spwntch/workspace-generator/commit/87ed7ea))
+- **next-storybook:** add storybook configuration ([1623322](https://github.com/spwntch/workspace-generator/commit/1623322))
+- **next-storybook:** add a fist story, for BrandProvider ([5d04d7b](https://github.com/spwntch/workspace-generator/commit/5d04d7b))
+- **next-storybook:** add spawntech brand to next-storybook ([db6ef91](https://github.com/spwntch/workspace-generator/commit/db6ef91))
+- **next-storybook:** add a fist story, for BrandProvider ([179ec57](https://github.com/spwntch/workspace-generator/commit/179ec57))
+- **next-storybook:** add spawntech brand to next-storybook ([1723f7f](https://github.com/spwntch/workspace-generator/commit/1723f7f))
+- **react-storybook:** export doc-templates from lib ([1a16e32](https://github.com/spwntch/workspace-generator/commit/1a16e32))
+- **react-storybook:** export doc-templates from lib ([dd3ee85](https://github.com/spwntch/workspace-generator/commit/dd3ee85))
+
+### ❤️  Thank You
+
+- zpydee @zpydee
+
+## 0.28.5 (2024-04-07)
+
+
+### 🚀 Features
+
+- **workspace-generator:** add tailwind capability to app ([d6c5e0d](https://github.com/spwntch/workspace-generator/commit/d6c5e0d))
+- **workspace-generator:** add spwntch deps and import spwntch styles ([c3931c5](https://github.com/spwntch/workspace-generator/commit/c3931c5))
+
+### 🩹 Fixes
+
+- **workspace-generator:** fix typos in generated files ([3dd5fe8](https://github.com/spwntch/workspace-generator/commit/3dd5fe8))
+- **workspace-generator:** comment out Providers to get app working ([34cf1c9](https://github.com/spwntch/workspace-generator/commit/34cf1c9))
+- **workspace-generator:** add Provider back into Layout ([b2e290c](https://github.com/spwntch/workspace-generator/commit/b2e290c))
+
+### ❤️  Thank You
+
+- zpydee @zpydee
+
 ## 0.28.4 (2024-04-06)